(mod) add BobinyLoader hooking mechanism
Add a simplistic global function hooking mechanism, exposed through bobiny.preHook or bobiny.postHook.

The hooks need a global function name and an interceptor function.
The pre-hook interceptor function receives arguments, and returns new arguments to be passed to the function.
The post-hook interceptor function receives the return value, and a new value to be passed to the caller.