(mod) add BobinyLoader override mechanism
Add a simplistic global function override mechanism, exposed through bobiny.override.

The override needs a global function name and an interceptor function.
The override interceptor function receives a delegate function as the first argument, and the arguments of the overridden function as the rest of the arguments. It may or may not call that delegate function with these arguments, and handle the result.