(mod) discover mods from filesystem
Add the ability to discover and load mods dynamically from the file system.

The folder BobinyMods is scanned for Lua files starting with `mod_`, and requires the files:

These mod files are expected to return a table with a function called `load`, which takes a path as an argument.
That path will refer to the BobinyMods folder ending with a slash, and can be used by the mod to require more files.

The loader will then call the `load(path)` function on every of them.