(mod) make bobiny-loader testable
It was hard to test bobiny-loader because it was required within the game files.
For this reason the bobiny-loader was replaced by a test stub if it was loaded from starting tests.

The test stub now acts as a proxy to the concrete implementation if a global toggle is set.
In addition, dependencies are not loaded automatically into the concrete implementation if run as part of tests.