(mod) identify and optionally call onAllModsLoaded when done loading
When all mods are loaded, call onAllModsLoaded on all mods that have such a function.
The argument to this function is object containing the mod descriptors of all loaded mods, including a reference to the mod itself.
This allows mods to call functions from other mods.

For this purpose, all mods now require a `identifier` string member, which will be present in the mod descriptors.