(mod) provide empty arrays on all objects
Since Lua tables are both used to represent array-like data structures or object-like data structures, an empty table cannot be immediately interpreted as either an empty array nor an empty object.

Therefore, arrays are always provided as a member of an object. In this case, when an table is empty or a table is interpreted as an object, also provide an empty array so that receivers have an easier time handling this case.